feat(@vtmn/svelte): remove animationDisabled, refactor tests
